Message channel over c-lightning
This commit creates a new MessageChannel
implementation `LNOnionMessageChannel`,
in `lnonion.py`, as well as classes
represeting LN peers and LN onionmessages.
In addition it implements a c-lightning
plugin `jmcl.py`, which forwards messages
received from other LN peers via the
`sendonionmessage` RPC (this is still an
experimental feature in c-lightning).
The messagechannel object must be configured
with a lightning-rpc socket file location,
so that it can send out these messages via
its c-lightning node.
Communcation between the plugin and jmdaemon
is accomplished using a twisted basic.LineReceiver
protocol.
An automatic installation of c-lightning as part
of Joinmarket's installed script is added, which
is activated with the flag --with-ln-messaging.
This instance of c-lightning is auto-started.
Documentation on this feature is added in
docs/lightning-message-channel.md.
Also adds basic lnonion module tests.