Switch to sendcustommsg
After 0.10.2 switched the onionmessage format,
and since we don't yet have a use for these messages
outside of a direct peer connection, and since the
onionmessage format is significantly more complex,
we switch for the time being to the hook and RPC
call for custommsg, instead.
Refactored the plugin to use a pure socket connection
instead of twisted, this avoids confusion since the plugin
fundamentally is blocking on iterating over stdin input, it
also avoids needing a big dependency like twisted when running
the plugin in whatever environment c-lightning is running in.
Secondly, added state management of peers and the requirement
for non-directory peers to handshake with directory peers on
startup, to ensure version match and pass basic info.
Adds config option to use non-bundled c-lightning.
(See comments on clightning-location var in the
default config in configure.py.)
Also modifies the docs and adds a note in the installation
script that including the bundled c-lightning is only required
if you do not intend to use your pre-existing
installation.
Cleans jmcl.py of any joinmarket codebase dependency.