End to end test; change handshake
The handshake is changed to include nick
and connection information.
Then we remove the getpeers request from non-dir
nodes, and instead the dir-nodes forward connection
info to clients whenever they get a privmsg from
another client. This avoids too much messaging spam
coming from the directory.
Also, a new test construction is found in
test/e2e-coinjoin-test.py; here, we automate not only the
setup and start of yieldgenerators in regtest, but
also we start up a JMWalletDaemon instance for the
Taker, who then runs a coinjoin with the docoinjoin
request.