Allow wallet lock in RPC-API without auth.
Fixes #1121. Prior to this commit, if a user
lost access to the authentication token for
a session with the wallet RPC, they would not
be able to lock the existing wallet before
authenticating again (getting a new token for
the same wallet, or a different one).
After this commit, the lock endpoint is no
longer authenticated, but will automatically
destroy the session for whatever wallet is
currently unlocked.