Accept fidelity bonds with ascii cert messages
Previously the `cert_msg` in the fidelity bond protocol involves a
binary public key. This is additional unnecessary complexity given
that we want other wallets to also implement the fidelity bond
protocol (see issue #993 for full discussion). So its good to move
to a certificate message which is entirely ascii.

This commit has ascii certficiate messages also be accepted as valid
along with the old certificate messages.